move part to components directory
